+++ scout/server/blueprints/variants/controllers.py
@@ -732,7 +732,7 @@ def _compound_follow_filter_clnsig(compound, compound_var_obj, query_form):
 
     There are some filter options that are rather unique, like the ClinVar one.
 
-    If clinsig_confident_always_returned is checked, variants are currently never dismissed on ClinSig alone.
+    If prioritise_clinvar is checked, variants are currently never dismissed on ClinSig alone.
 
     Args:
         compound(dict)
@@ -743,8 +743,7 @@ def _compound_follow_filter_clnsig(compound, compound_var_obj, query_form):
     query_rank = []
     query_str_rank = []
 
-    clinsig_always_returned = query_form.get("clinsig_confident_always_returned")
-    if clinsig_always_returned:
+    if query_form.get("prioritise_clinvar"):
         return False
 
     clinsig = query_form.get("clinsig")